# Significantly improve whisper.cpp inference quality