**We've wrapped up our analysis comparing the `log_mel_spectrogram` generation between whisper.cpp and OpenAI's Whisper.**

### To summarize the main issues we found in whisper.cpp:

1. The Stage-1 padding (zero padding) is inadequate. While OpenAI's Whisper uses a padding of `480,000` samples, whisper.cpp only goes between `240,000` and `480,000`.

2. Stage-2 (reflective padding) is missing. This oversight can introduce edge effects when whisper.cpp processes the STFT, potentially causing spectral leakage.

3. The frame count calculation isn't accurate.

4. After performing the FFT, it mistakenly aggregates the amplitudes from symmetrical frequency bins.

### On top of these, whisper.cpp presents a couple of secondary concerns:

1. The trig functions in the C++ <cmath> library default to FP64 computations, in contrast to PyTorch's FP32. This leads to significant discrepancies, especially with smaller angles.

2. While whisper.cpp offers a feature to shift from mono to a kind of simulated stereo, it's uncertain if the model fully supports stereo audio.

3. whisper.cpp doesn't remove the last frame like OpenAI's Whisper does, which could lead to some potential issues.

### With these findings in hand, we're set to fix whisper.cpp.

> Uncovering these issues wouldn't have been possible without the collective support and assistance from everyone. A heartfelt thanks to @regularfry for his invaluable technical expertise in signal processing and for suggesting solutions. I'm grateful to @gauvainjl for introducing the WER testing method. A special shoutout to @ggerganov for his unwavering support. And last but not least, a big thank you to the entire community for their continued attention to this project.

[Full report](https://github.com/ggerganov/whisper.cpp/discussions/1163)


![Whisper_FFT](https://github.com/ggerganov/whisper.cpp/assets/129547291/7c75d27b-d475-48d1-b713-449db299fdf6)

> 

### After conducting some experiments, I found that the smaller the model size, the more significant the improvement in inference quality. There is almost no noticeable difference when using the `large` model.

`diffusion2023-07-03.wav` `ggml-model-tiny.bin`
**A complete comparison is [here](https://www.diffchecker.com/IBr5U0hF/).**

<img width="1096" alt="image" src="https://github.com/ggerganov/whisper.cpp/assets/129547291/0403fe15-39bd-42a4-a404-4e2db3719b2e">

> 

`diffusion2023-07-03.wav` `ggml-model-base.bin`
**A complete comparison is [here](https://www.diffchecker.com/ypGwLo7J/).**

<img width="1110" alt="image" src="https://github.com/ggerganov/whisper.cpp/assets/129547291/98ff6f3a-c36b-40ac-9d2d-0e6dd5b5b1e8">

> 

`a13.wav`
**A complete comparison is currently unavailable.**

<img width="1121" alt="image" src="https://github.com/ggerganov/whisper.cpp/assets/129547291/e78c2f00-a9c9-40b2-928f-16739775527b">