@bobqianic 2023-08-03
Interestingly, when the Speed-up mode is switched on – and it's worth noting that it's off by default – the system occasionally generates numerous duplicate sentences. The cause of this issue remains unclear at the moment, but rest assured, I'm working diligently to resolve it as quickly as possible.

<img width="1096" alt="image" src="https://github.com/ggerganov/whisper.cpp/assets/129547291/f5b1a338-06e7-41f0-a709-555d081e33bb">

Significant quality improvement `speedup mode ON`

> 
### What's that?
<img width="567" alt="image" src="https://github.com/ggerganov/whisper.cpp/assets/129547291/0dcd369a-41ca-410a-8f1c-ea2ba8d9b2e4">