@bobqianic 2023-08-03
I believe the most effective solution to address these issues is a complete rewrite of the `log_mel_spectrogram function` in whisper.cpp. We should do this in line with the guidelines provided in `audio.py` and the `Torch documentation`. Generating a correct `log_mel_spectrogram` is crucial as the Whisper model uses this as its input.