@bobqianic 2023-08-04
I would like to share some exciting news with you all. Now, in the `Log_Mel` function, when `FFT_IN` is all zeros, the FFT will not be computed, and the result will be output directly, significantly accelerating the calculation speed of the `Log_Mel_Spectrogram`. On the basis of the `Sin/Cos_Cache optimization` in #1142, the calculation time has been reduced by approximately **`42%`**, which is a huge improvement! @ggerganov @AlexandrGraschenkov

`Ryzen 7 5700X` `a13.wav`
| Original (ms) | Sin/Cos_Cache (ms) | Sin/Cos_Cache + Log_Mel_opt_v1 (ms) | Sin/Cos_Cache + Log_Mel_opt_v2 (ms) |
| :-----------: | :-----------: | :-----------: | :-----------: |
| 652.68 | 190.68 | 101.16 | 94.24 |
| 640.01 | 175.93 | 92.70 | 98.43 |
| 647.10 | 168.24 | 103.22 | 92.45 |
| 649.84 | 172.56 | 105.32 | 93.28 |
| 637.95 | 176.11 | 94.04 | 100.21 |
| 637.75 | 176.59 | 106.42 | 95.28 |
| 635.10 | 171.02 | 95.02 | 93.74 |
| 632.21 | 169.23 | 98.07 | 93.02 |
| 635.55 | 167.07 | 105.57 | 98.53 |
| 635.26 | 168.94 | 109.16 | 93.75 |
| 645.66 | 170.45 | 110.75 | 94.10 |

> 

`Ryzen 7 5700X` `diffusion2023-07-03.wav`
| Original (ms) | Sin/Cos_Cache (ms) | Sin/Cos_Cache + Log_Mel_opt_v1 (ms) | Sin/Cos_Cache + Log_Mel_opt_v2 (ms) |
| :-----------: | :-----------: | :-----------: | :-----------: |
| 5609.77 | 1478.99 | 859.58 | 857.02 |
| 5630.08 | 1467.94 | 859.60 | 844.49 |
| 5662.63 | 1467.00 | 849.40 | 858.90 |