@vadi2 2023-08-04
Excellent work!

To get a better perspective, how does it affect the overall transcription speed from start to end?